rag-processor/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-astra
PyPDF2
python-dotenv
